examples/chemcrow/requirements.txt
# ChemCrow-HPC Requirements
# High-performance chemical analysis with Flowgentic

# Core dependencies
flowgentic>=0.1.0
rdkit-pypi>=2023.9.1
pandas>=2.0.0
numpy>=1.24.0

# LangChain and LLM integration
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40

# AsyncFlow and HPC
radical-asyncflow>=0.1.0
academy-py>=0.1.0

# Chemistry and scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0

# API and web services
aiohttp>=3.8.0
httpx>=0.24.0
requests>=2.31.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Optional dependencies for advanced features
molbloom>=0.3.0          # Bloom filters for chemical databases
paper-qa>=4.0.0            # Scientific paper analysis
paperscraper>=0.2.0        # Academic paper scraping

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# HPC and cluster computing
radical-saga>=1.24.0       # SAGA-based job management
radical-pilot>=1.24.0      # Pilot-based execution
radical-utils>=1.24.0      # Utility functions